app/requirements_updated.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.12

# Gemini AI
google-generativeai>=0.8.0

# Database
sqlalchemy>=2.0.36
pymysql>=1.1.1

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Environment
python-dotenv>=1.0.1

# Utilities
httpx>=0.27.0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
